Rename NewDropdown component to Dropdown.
