+++ resources/js/Partials/Items/ItemList.vue
@@ -13,7 +13,7 @@
                 </div>
                 <div slot="table-actions" class="flex items-center justify-between">
                     <span class="text-blue-500 text-sm font-semibold leading-loose mr-2 inline-block mt-tenth cursor-pointer" @click="clearSearch()">Clear</span>
-                    <new-dropdown v-if="$page.auth.user.can.administerItems" class="mr-1" right="10" width="180">
+                    <dropdown v-if="$page.auth.user.can.administerItems" class="mr-1" right="10" width="180">
                         <div slot="trigger" class="flex items-center cursor-pointer select-none group">
                             <div class="text-blue-900 group-hover:text-blue-700 focus:text-blue-700 mr-1 whitespace-no-wrap">
                                 <span class="inline text-gray-800 text-sm font-semibold">Options</span>
@@ -59,7 +59,7 @@ import Book from '@/Shared/Icons/Book';
 import Plus from '@/Shared/Icons/Plus';
 import IconBase from '@/Shared/IconBase';
 import Checkbox from '@/Shared/Checkbox';
-import NewDropdown from '@/Shared/NewDropdown';
+import Dropdown from '@/Shared/Dropdown';
 import { VueGoodTable } from 'vue-good-table';
 import CheveronDown from '@/Shared/Icons/CheveronDown';
 import WatchesForErrors from 'Mixins/WatchesForErrors';
@@ -70,7 +70,7 @@ export default {
         Plus,
         IconBase,
         Checkbox,
-        NewDropdown,
+        Dropdown,
         CheveronDown,
         VueGoodTable,
     },
