+++ resources/js/Partials/Materials/Fabrics/FabricsTable.vue
@@ -14,7 +14,7 @@
         <template slot="table-actions">
             <div class="flex">
                 <span class="text-blue-500 text-sm font-semibold leading-loose mr-2 inline-block mt-tenth cursor-pointer" @click="clearSearch()">Clear</span>
-                <new-dropdown class="mt-1 mr-1" right="10" width="180">
+                <dropdown class="mt-1 mr-1" right="10" width="180">
                     <div slot="trigger" class="flex items-center cursor-pointer select-none group">
                         <div class="text-blue-900 group-hover:text-blue-700 focus:text-blue-700 mr-1 whitespace-no-wrap">
                             <span class="inline text-sm">Options</span>
@@ -53,7 +53,7 @@ import IconBase from '@/Shared/IconBase';
 import Checkbox from '@/Shared/Checkbox';
 import Printer from '@/Shared/Icons/Printer';
 import { VueGoodTable } from 'vue-good-table';
-import NewDropdown from '@/Shared/NewDropdown';
+import Dropdown from '@/Shared/Dropdown';
 import CheveronDown from '@/Shared/Icons/CheveronDown';
 
 export default {
@@ -62,7 +62,7 @@ export default {
         Printer,
         IconBase,
         Checkbox,
-        NewDropdown,
+        Dropdown,
         VueGoodTable,
         CheveronDown,
     },
