+++ resources/js/Partials/Materials/Printers/PrintersTable.vue
@@ -20,7 +20,7 @@
         <template slot="table-actions">
             <div class="flex justify-between items-center">
                 <span class="text-blue-500 text-sm font-semibold leading-loose mr-2 inline-block mt-tenth cursor-pointer" @click="clearSearch()">Clear</span>
-                <new-dropdown class="mr-1" right="10" width="180">
+                <dropdown class="mr-1" right="10" width="180">
                     <div slot="trigger" class="flex items-center cursor-pointer select-none group">
                         <div class="text-blue-900 group-hover:text-blue-700 focus:text-blue-700 mr-1 whitespace-no-wrap">
                             <span class="inline text-sm">Options</span>
@@ -52,7 +52,7 @@ import Plus from '@/Shared/Icons/Plus';
 import IconBase from '@/Shared/IconBase';
 import Checkbox from '@/Shared/Checkbox';
 import { VueGoodTable } from 'vue-good-table';
-import NewDropdown from '@/Shared/NewDropdown';
+import Dropdown from '@/Shared/Dropdown';
 import CheveronDown from '@/Shared/Icons/CheveronDown';
 
 export default {
@@ -60,7 +60,7 @@ export default {
         Plus,
         IconBase,
         Checkbox,
-        NewDropdown,
+        Dropdown,
         VueGoodTable,
         CheveronDown,
     },
