+++ resources/js/Shared/SearchFilter.vue
@@ -1,7 +1,7 @@
 <template>
     <div class="flex items-center">
         <div class="flex w-full bg-white shadow rounded">
-            <new-dropdown class="px-4 md:px-6 rounded-l border-r hover:bg-gray-100 focus:border-white focus:z-10" right="10">
+            <dropdown class="px-4 md:px-6 rounded-l border-r hover:bg-gray-100 focus:border-white focus:z-10" right="10">
                 <div slot="trigger" class="flex items-baseline">
                     <span class="text-gray-900 hidden md:inline">Filter</span>
                     <svg class="w-2 h-2 fill-gray-800 md:ml-2" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 961.243 599.998">
@@ -19,11 +19,11 @@
 </template>
 
 <script>
-import NewDropdown from '@/Shared/NewDropdown';
+import Dropdown from '@/Shared/Dropdown';
 
 export default {
     components: {
-        NewDropdown,
+        Dropdown,
     },
     props: {
         value: String,
